# 5x Faster Set Intersections in AVX-512 & SVE2